src/main/java/leetcode/Problem219.java
package leetcode;

import java.util.HashSet;
import java.util.Set;

public class Problem219 {
    public boolean containsNearbyDuplicate(int[] nums, int k) {
        if (nums.length < 2) {
            return false;
        }
        Set<Integer> set = new HashSet<>(k, 1);
        for (int i = 0; i < nums.length; i++) {
            if (!set.add(nums[i])) {
                return true;
            }
            if (i >= k) {
                set.remove(nums[i - k]);
            }
        }
        return false;
    }
}
